messaging_app/requirements.txt
asgiref==3.9.1
distlib==0.3.8
Django==5.2.4
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
filelock==3.13.1
pbr==6.0.0
platformdirs==4.2.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
stevedore==5.2.0
tzdata==2025.2
virtualenv==20.25.1
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.0
djangorestframework-simplejwt==5.3.0
drf-nested-routers==0.94.1
mysqlclient==2.2.0
